Transform RawLocks into a map of interfaces
RawLocks are now decoded in two steps:

1. by the builder package, at the same time as the raw config. It's
transformed into a map of interfaces.
2. then by specialized defkinds. It's then transformed into specific
DefinitionLocks.

This seems a bit cleaner than before since there's now some form of
symmetry between RawConfig and RawLocks. Moreover as the locks are
decoded by the builder, this package is now able to arbitrarily modify
them which is going to help refactorize the way webserver definitions
and locks are embedded into other definitions.