Add tools to help review changes on LLB DAGs dumps
This commit adds two new tools. The first one takes two graphs, compare
them and color in grey all the nodes that aren't in both graphs.

The second tool use the first one to open up a colored graph for every
dumps that changed in a commit range (or unstaged dumps if not range is
provided).