Add support for npm
So far, nodejs handler was only using yarn to install packages. This
commit implements npm support by checking if package-lock.json is in the
source context. If so, NPM is used to build, otherwise yarn is used.

A new method FileExists() has been added to StateSolver. However, the
current implementation isn't optimal as it tries to read the whole file
to determine if it exists. This works but it should definitively
improved.