Use a map to associate pkgsolver Types with impls
So far, only the APT package solver was used to pin package versions.
This commit introduces pkgsolver.PackageSolversMap which is used to
associate pkgsolver.SolverType with factory functions.

This change is needed to properly implement Alpine support in a later
commit.

Unfortunately it's not possible to transform these PackageSolvers into
plain functions implementing a PackageSolver.ResolveVersions() signature
because PackageSolver interface is mocked by gomock to ease testing.