Ignore empty platform list from composer.lock files
So far, zbuild was expecting composer.lock files to contain a "platform"
parameter with some extension requirements listed. However, on the
current project I'm trying to use zbuild on, these parameters are empty.

This leads to an error when these files are decoded by zbuild, because
it expects a map[string]string whereas an empty PHP associative array is
represented as an empty list (eg. `[]`) when converted to JSON.