Support --no-cache CLI flag
When --no-cache flag is passed to `docker buildx build` or `buildctl
build`, Buildkit passes a "no-cache" option. It's then the
responsibility of the Buildkit frontend to mark each and every Run/Copy
step with llb.IgnoreCache constraint.

This commit introduces a new field to builddef.BuildOpts named
IgnoreCache. The generic builder package takes care of setting this
value when the "no-cache" option is received. It's then the
responsibility of specialized kind handlers to use this value correctly.