Use the source instead of the dest to populate llb.includepattern
The last commit introduced a bug in the webserver builder. This bug is
preventing it to copy config files when they originate from a subdir of
the build context. Moreover, the path used to copy files was prefixed
with the build context subdir but the path used by llb.includepattern
wasn't.