Don't check changes on go.sum on TravisCI
This CI job runs all the `make gen-*` commands and then check if
anything has changed in the repo. This is used to ensure that testdata
files and doc diagrams are always up-to-date. However, for some unknown
reason, Go might add some hashes to go.sum on the CI whereas it don't
want to add them from local dev machine.

As such, the check-generated-files script has been changed to always
discard changes on go.sum before checking the status of the repo.