Limit the context used by FileExists to the path to check
So far, the FileExists method of statesolver.Buildkit is based on
reference.ReadFile() method, coming from Buildkit SDK. This method was
called against a reference based on the source/build context passed as
argument. However, the referenced state was including the whole context.
Because of that, Buildkit was showing different context size from one
build to another.

Also, note that it'd be better to stat the path instead of reading it.
However, this commit led to discovering a bug in Buildkit: stating a
nonexistant file makes the frontend fail with an error message (whereas
the `ReadFile()` method is properly returning an os.IsNotExist error).